backend/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-postgres>=0.0.17",
    "pgvector>=0.3.6",
    "protobuf>=7.34.0",
    "psycopg[binary]>=3.3.3",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
